Projection: Move the version schema inclusion into a separate module
